requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.1.7
pillow==11.1.0
psycopg==3.2.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
